// packages/text/types.ts
import { ReactNode } from 'react';
import { TextStyle } from 'react-native';
import { MarginProps, PaddingProps } from '../utils';

export interface TextProps extends MarginProps, PaddingProps {
  children?: ReactNode;
  css?: TextStyle;
  style?: TextStyle;
  fontSize?: number;
  fontFamily?: string;
  color?: string;
  fontWeight?:
    | 'normal'
    | 'bold'
    | '100'
    | '200'
    | '300'
    | '400'
    | '500'
    | '600'
    | '700'
    | '800'
    | '900';
  textAlign?: 'left' | 'center' | 'right' | 'justify';
  textDecoration?: 'none' | 'underline' | 'line-through' | 'underline line-through';
  textTransform?: 'none' | 'capitalize' | 'uppercase' | 'lowercase';
  fontStyle?: 'normal' | 'italic';
  lineHeight?: number;
  letterSpacing?: number;
  numberOfLines?: number;
  ellipsizeMode?: 'head' | 'middle' | 'tail' | 'clip';
}